requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.0
azure-eventhub==5.11.5
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.8.0
azure-keyvault-secrets==4.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
frozenlist==1.4.1
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
